widgets fields refactor - file -> package
