+++ BaCa2/widgets/forms/__init__.py
@@ -1,2 +1,1 @@
-from widgets.forms.base import BaCa2Form, FormWidget
-from widgets.forms.fields import get_field_validation_status
+from .base import BaCa2Form, FormWidget
